Extract mouse picking into Input::Selection
Move the cursor-picking functions (SelectObjects and its internal helpers
SelectItem/SelectCharacter/SelectOperate, ~360 lines) out of ZzzInterface.cpp
into Input/Selection.{h,cpp} under namespace Input::Selection. Only SelectObjects
is public (two scene call sites); the others were already file-local. The
Selected* globals it writes stay in ZzzInterface.cpp, declared via extern for now.
No behaviour change.
