+++ src/source/Scenes/CharacterScene.cpp
@@ -12,6 +12,7 @@
 #include "Engine/Object/ZzzCharacter.h"
 #include "Render/Terrain/ZzzLodTerrain.h"
 #include "Engine/Object/ZzzInterface.h"
+#include "Input/Selection.h"
 #include "Engine/Object/ZzzInfomation.h"
 #include "Render/Effects/ZzzEffect.h"
 #include "Engine/AI/GOBoid.h"
@@ -297,7 +298,7 @@ static void RenderCharacterScene3D()
     RenderCharactersClient();
 
     if (!CUIMng::Instance().IsCursorOnUI())
-        SelectObjects();
+        Input::Selection::SelectObjects();
 
     RenderMount();
     RenderBlurs();
